Extract Expr nodes to expr.py, slim symbolic.py to drop-ins only
New module structure:
  expr.py     (280 lines) — Expr base, all node classes, coercion
  symbolic.py (149 lines) — drop-in functions, sym(), re-exports
  simplify.py (241 lines) — algebraic rewrite rules

symbolic.py went from 860 → 149 lines. All Expr nodes re-exported
from symbolic.py for backward compatibility.

render.py, latex_build.py, simplify.py now import from ga.expr directly.
